+++ build.dev.js
@@ -2,10 +2,16 @@
 
 require('shelljs/global');
 
+const env       = 'dev';
+
+
 const fs        = require('fs');
-const utils     = require('./build.utils.js');
 const chokidar  = require('chokidar');
 const sass      = require('node-sass');
+const utils     = require('./build.utils.js');
+const postcss   = require('./postcss.'+env+'.js');
+
+/* References to shared tools are found in build.utils.js */
 
 const console   = utils.console;
 const colors    = utils.colors;
@@ -15,13 +21,14 @@ const log       = utils.log;
 const warn      = utils.warn;
 const clean     = utils.clean;
 
-const env = 'dev';
-
 let canWatch = true;
 let isCompiling = false;
 let hasInit = false;
 let styleFiles = [];
 let hasCompletedFirstStylePass = false;
+let postcssConfig = ' -u';
+
+/* Test for arguments the ngr cli spits out */
 
 process.argv.forEach((arg)=>{
   if (arg.includes('watch')) {
@@ -29,7 +36,22 @@ process.argv.forEach((arg)=>{
   }
 });
 
-/* Copy */
+/* Process PostCSS CLI plugins for the --use argument */
+
+for (let cssProp in postcss.plugins) {
+  postcssConfig += ' '+cssProp;
+}
+
+/*
+
+  Copy Tasks
+
+- public: Copies the contents of the src/public folder
+- file: Copies a file to /build
+- html: Copies all .html files to /build
+- lib: Copies files and folders from /node_modules to /build/lib
+
+*/
 
 const copy = {
     public: (path) => {
@@ -79,34 +101,15 @@ const copy = {
     }
 };
 
-/* Compile */
+/*
+
+  Compile Tasks
+
+- ts: Compiles Typescript files from /src into /build
+
+*/
 
 const compile = {
-
-    clean: (path) => {
-
-      const outFile = path ? path : './build/bundle.js';
-
-      fs.readFile(outFile, 'utf8', function(err, contents) {
-        if(!err) {
-            contents = contents.replace(utils.multilineComment, '');
-            contents = contents.replace(utils.singleLineComment, '');
-            fs.writeFile(outFile, contents, function(err){
-              if(!err) {
-              //  log('Cleaned up', 'comments', 'from', outFile);
-              } else {
-                warn(err);
-              }
-            });
-        } else {
-          warn(err);
-        }
-
-      });
-
-
-    },
-
 
     ts : (path) => {
 
@@ -141,7 +144,21 @@ const compile = {
     }
 };
 
-/* Styling */
+/*
+
+  Style Tasks
+
+  - file: Styles a single file.
+         - If the file is in the /src/styles folder it will compile /src/styles/style.scss
+         - If the file is elsewhere, like part of a Component, it will compile into the
+          appropriate folder in the /build directory
+  - src: Compiles the global styles
+
+  SASS render method is called and fs writes the files to appropriate folder
+  PostCSS processes the file in place, using the --replace argument
+
+
+*/
 
 let style = {
 
@@ -166,7 +183,7 @@ let style = {
 
                 if (watch === true) log('node-sass', 'compiled', 'component style at', outFile);
 
-                let postcss = exec('postcss -c postcss.'+env+'.js -r '+outFile, function(code, output, error) {
+                let postcss = exec('postcss ./'+outFile+' -c ./postcss.'+env+'.js -r'+postcssConfig, function() {
 
                     if ( (styleFiles.indexOf(path) === styleFiles.length - 1) && hasCompletedFirstStylePass === false) {
                       log('libsass and postcss', 'compiled', 'for', colors.bold(colors.cyan(env)));
@@ -210,22 +227,36 @@ let style = {
     }
 };
 
-/* Init */
+/*
+
+  Init Tasks
+
+  A sequence of commands needed to clean and start the dev build
+
+*/
 
 let init = function() {
 
     rm('-rf', paths.rootDir+'/.tmp/');
     rm('-rf', './'+paths.build);
     rm('-rf', './ngfactory');
+
     mkdir('./'+paths.build);
     mkdir('./'+paths.build+'/lib');
+
     copy.lib();
     copy.public();
     compile.ts();
 
 };
 
-/* Watcher */
+/*
+
+  Watcher
+
+  Chokidar is used to watch files, run the above methods
+
+*/
 
 
 let watcher = chokidar.watch('./'+paths.src+'/**/*.*', {