+++ postcss.dev.js
@@ -1,7 +1,7 @@
 module.exports = {
   plugins: {
     'autoprefixer': {
-        'browsers': '> 5%'
+        browsers: 'last 20 versions'
     }
   }
 }